% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic_sp2.R
\name{.fit_ic_sp}
\alias{.fit_ic_sp}
\title{Fitter function for interval censored semi-parametric models}
\usage{
.fit_ic_sp(x, y, weights, strata, model_type, other_info)
}
\arguments{
\item{x}{Covariate matrix}

\item{y}{Response matrix from \code{Surv(..., type = "interval2")}}

\item{weights}{Observation weights}

\item{strata}{Strata factor}

\item{model_type}{Model type: "ic_ph" or "ic_po"}

\item{other_info}{List of other fitting options}
}
\value{
Fitted model object
}
\description{
Fitter function for interval censored semi-parametric models
}
\details{
For advanced use only. This function is called internally by \code{ic_sp_ph}
and \code{ic_sp_po}.
}
