% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic_sp2.R
\name{ic_sp_settings}
\alias{ic_sp_settings}
\title{Settings for ic_sp_ph}
\usage{
ic_sp_settings(
  useGA = TRUE,
  maxIter = 10000,
  baseUpdates = 5,
  regStart = NULL,
  derivMethod = c(12, 1),
  updateCovars = TRUE
)
}
\arguments{
\item{useGA}{Should constrained gradient ascent step be used?}

\item{maxIter}{Maximum iterations}

\item{baseUpdates}{number of baseline updates (ICM + GA) per iteration}

\item{regStart}{Initial values for regression parameters}

\item{derivMethod}{Method for derivative calculations.}

\item{updateCovars}{Should covariates be updated during fitting?

@description
Creates the control options for the \code{ic_sp} function.
Defaults not intended to be changed for use in standard analyses.}
}
\description{
Settings for ic_sp_ph
}
\details{
The constrained gradient step, actived by \code{useGA = T},
is a step that was added to improve the convergence in a special case.
The option to turn it off is only in place to help demonstrate it's utility.

\code{regStart} also for seeding of initial value of regression parameters. Intended for use in ``warm start" for bootstrap samples
and providing fixed regression parameters when calculating fit in qq-plots.
}
