% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IR_diabetes}
\alias{IR_diabetes}
\title{Interval censored time from diabetes onset to diabetic nephronpathy}
\format{
An object of class \code{data.frame} with 731 rows and 3 columns.
}
\usage{
IR_diabetes
}
\description{
Data set contains interval censored survival time for time from onset of
diabetes to to diabetic nephronpathy. Identical to the \code{diabetes}
dataset found in the package \code{glrt}.
}
\section{Fields}{

\describe{
\item{\code{left}}{left side of observation interval}

\item{\code{right}}{right side of observation interval}

\item{\code{gender}}{gender of subject}
}}

\examples{
 head(IR_diabetes)
 fit <- ic_par(cbind(left, right) ~ gender, 
               data = IR_diabetes,
               model = "po",
               dist = "loglogistic")
}
\references{
Borch-Johnsens, K, Andersen, P and Decker, T (1985).
"The effect of proteinuria on relative mortality in Type I (insulin-dependent) diabetes mellitus."
Diabetologia, 28, 590-596.
}
\keyword{datasets}
