% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{lines.ic_sp2}
\alias{lines.ic_sp2}
\title{Lines method for ic_sp2 objects}
\usage{
\method{lines}{ic_sp2}(x, y, type = "s", strata, ...)
}
\arguments{
\item{x}{an object of class 'ic_sp2'}

\item{y}{ignored}

\item{type}{the type of plot to produce, default is "l" for lines}

\item{strata}{Optional vector of indices to subset the strata to be plotted.}

\item{...}{additional arguments passed to the lines function}
}
\description{
Lines method for ic_sp2 objects
}
\details{
Due to the interval censoring, the survival curves are not unique,
so two lines are drawn for each curve: one for the lower bound and
one for the upper bound of the interval.
A pair of lines is drawn for each stratum, if strata are present in the model.
Therefore to unambiguously specify graphics parameters such as colour or
line type, the user should specify them as a vector of length equal to the
number of strata (for strata specific pars) or twice the number of strata
(for line specific pars, e.g. strata 1 lower, strata 1 upper, strata 2 lower,
strata 2 upper).
}
