% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_utilities.R
\name{predict.icsp2}
\alias{predict.icsp2}
\title{Predictions from icenReg Regression Model}
\usage{
\method{predict}{icsp2}(object, type = "response", newdata = NULL, ...)
}
\arguments{
\item{object}{Model fit with \code{ic_par} or \code{ic_sp}}

\item{type}{type of prediction. Options include \code{"lp", "response"}}

\item{newdata}{\code{data.frame} containing covariates}

\item{...}{other arguments (will be ignored)}
}
\description{
Gets various estimates from an \code{ic_np}, \code{ic_sp} or \code{ic_par} object.
}
\details{
If \code{newdata} is left blank, will provide estimates for original data set.

For the argument \code{type}, there are two options. \code{"lp"} provides the
linear predictor for each subject (i.e. in a proportional hazards model,
this is the log-hazards ratio, in proportional odds, the log proporitonal odds),
\code{"response"} provides the median response value for each subject,
\emph{conditional on that subject's covariates, but ignoring their actual response interval}.
Use \code{imputeCens} to impute the censored values.
}
\examples{
simdata <- simIC_weib(n = 500, b1 = .3, b2 = -.3,
                      inspections = 6,
                      inspectLength = 1)

fit <- ic_par(cbind(l, u) ~ x1 + x2,
              data = simdata)

imputedValues <- predict(fit)
}
\author{
Clifford Anderson-Bergman
}
