% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic_sp2.R
\name{print.ic_sp2}
\alias{print.ic_sp2}
\title{Print method for ic_sp2 objects}
\usage{
\method{print}{ic_sp2}(x, ...)
}
\arguments{
\item{x}{Fitted model object from \code{ic_sp}}

\item{...}{Additional arguments.}
}
\description{
Print method for ic_sp2 objects
}
