% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic_sp2.R
\name{profile_fit}
\alias{profile_fit}
\title{Refit the model with fixed regression parameters}
\usage{
profile_fit(object, beta = object$coefficients)
}
\arguments{
\item{object}{Fitted model object from \code{ic_sp}}

\item{beta}{Vector of regression parameters to fix in the refit. Default is the original fitted regression parameters.}
}
\value{
Raw fitted model object with fixed regression parameters.
}
\description{
Refit the model with fixed regression parameters
}
