% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic_sp.R
\name{profile_likelihood_covariance}
\alias{profile_likelihood_covariance}
\title{Profile Likelihood Covariance for Semi-Parametric Models}
\usage{
profile_likelihood_covariance(fit, constant = 1, ...)
}
\arguments{
\item{fit}{Fitted model object from \code{ic_sp}}

\item{constant}{Multiplier for the constant \code{h_n} in the profile likelihood.}

\item{...}{Additional arguments.}
}
\description{
Profile Likelihood Covariance for Semi-Parametric Models
}
