% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport.R
\name{reexports}
\alias{reexports}
\alias{Surv}
\alias{strata}
\title{Re-exported functions}
\usage{
Surv(
  time,
  time2,
  event,
  type = c("right", "left", "interval", "counting", "interval2"),
  origin = 0
)

strata(..., na.group = FALSE, shortlabel, sep = ", ")
}
\description{
Re-exported functions
}
