% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic_sp2.R
\name{vcov.ic_po}
\alias{vcov.ic_po}
\title{Profile Likelihood Covariance for Semi-Parametric Models}
\usage{
\method{vcov}{ic_po}(object, constant = 1, ...)
}
\arguments{
\item{constant}{Multiplier for the constant \code{h_n} in the profile likelihood.}

\item{...}{Additional arguments.}

\item{fit}{Fitted model object from \code{ic_sp}}
}
\description{
Profile Likelihood Covariance for Semi-Parametric Models
}
