% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp_logrank.R
\name{cpp_logrank}
\alias{cpp_logrank}
\title{Internal fast log-rank test implementation}
\usage{
cpp_logrank(time, event, group)
}
\arguments{
\item{time}{Numeric vector of survival/censoring times.}

\item{event}{Logical vector of event indicators (TRUE = event, FALSE = censored).}

\item{group}{Integer vector of group labels, coded as consecutive integers
(i.e. \verb{1, 2, ...}). This is not checked.}
}
\value{
An object of class \code{"cpp_logrank"}, a list with components:
\describe{
\item{observed}{Numeric vector of observed events per group.}
\item{expected}{Numeric vector of expected events per group.}
\item{variance}{Numeric variance matrix}
\item{n_by_group}{Numeric vector of sample sizes per group.}
}
}
\description{
This is an internal implementation of right-censored logrank test
for use within \link{ic_logrank} where many tests need to be performed
efficiently for variance calculation.
}
\details{
Compared to Survival::survdiff, this implementation is faster but does
not have the same flexibility or checks.
}
\examples{
set.seed(1992)
time  <- rexp(100)
event <- sample(c(TRUE, FALSE), 100, replace = TRUE)
group <- sample(1:2, 100, replace = TRUE)
cpp_logrank(time, event, group)

}
\keyword{internal}
