% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic_sp2.R
\name{ic_sp2}
\alias{ic_sp2}
\alias{ic_sp_ph}
\alias{ic_sp_po}
\title{Fit a semi-parametric model for interval-censored data}
\usage{
ic_sp2(
  formula,
  data,
  weights,
  subset,
  na.action,
  B = c(0, 1),
  control = ic_sp_control(...),
  model = c("ph", "po"),
  profile_ci = NULL,
  ...
)

ic_sp_ph(
  formula,
  data,
  weights,
  subset,
  na.action,
  B = c(0, 1),
  control = ic_sp_control(...),
  model = c("ph", "po"),
  profile_ci = NULL,
  ...
)

ic_sp_po(
  formula,
  data,
  weights,
  subset,
  na.action,
  B = c(0, 1),
  control = ic_sp_control(...),
  model = c("ph", "po"),
  profile_ci = NULL,
  ...
)
}
\arguments{
\item{formula}{A model formula with \code{Surv(l, u, type = 'interval2')} response and covariates on the right-hand side. May also contain \code{strata()} terms.}

\item{data}{A data frame containing the variables in the formula, including strata terms.}

\item{weights}{Optional vector of weights for each observation, or the name of a variable in \code{data} containing the weights.}

\item{subset}{Optional expression indicating a subset of the rows of \code{data} to be used in the fit.}

\item{na.action}{Optional function to handle missing data. Default is \code{na.omit}.}

\item{B}{A vector of length 2 giving the lower and upper bounds for the observation times. Default is c(0, 1).}

\item{control}{A list of control settings, with defaults created by \code{\link[=ic_sp_control]{ic_sp_control()}}.}

\item{model}{Type of model to fit. Choices are \code{"ph"} for proportional hazards and \code{"po"} for proportional odds. Default is \code{"ph"}.
This is normally determined by the function aliases \code{ic_sp_ph} and \code{ic_sp_po}.}

\item{profile_ci}{Confidence level for profile likelihood confidence intervals based on the likelihood ratio.
Set to \code{NULL} to skip this profile likelihood confidence interval calculations.}

\item{...}{Additional arguments passed to control.}
}
\value{
An object of class \code{"ic_sp2"}, a list containing the fitted model information, including coefficients, variance-covariance matrix, and other details.
}
\description{
This function fits a semi-parametric model for interval-censored data using
either the proportional hazards or proportional odds model. Aliases \code{ic_sp_ph()} and \code{ic_sp_po()} are provided for convenience.
}
\examples{
fit_ph <- ic_sp_ph(Surv(l, u, type = "interval2") ~ grp, data = miceData, profile_ci = 0.95)
summary(fit_ph)

# The CI for the regression parameter(s) based on the likelihood ratio:
fit_ph$profile_ci

confint(fit_ph, level = 0.95)
}
