% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic_sp2.R
\name{ic_sp2}
\alias{ic_sp2}
\alias{ic_sp_ph}
\alias{ic_sp_po}
\title{Fit a semi-parametric model for interval-censored data}
\usage{
ic_sp2(
  formula,
  data,
  weights,
  subset,
  na.action,
  B = c(0, 1),
  settings = ic_sp_settings(),
  model = c("ph", "po")
)

ic_sp_ph(
  formula,
  data,
  weights,
  subset,
  na.action,
  B = c(0, 1),
  settings = ic_sp_settings(),
  model = c("ph", "po")
)

ic_sp_po(
  formula,
  data,
  weights,
  subset,
  na.action,
  B = c(0, 1),
  settings = ic_sp_settings(),
  model = c("ph", "po")
)
}
\arguments{
\item{formula}{A model formula with Surv(l, u, type = 'interval2') response and covariates on the right-hand side. May also contain \code{strata()} terms.}

\item{data}{A data frame containing the variables in the formula, including strata terms.}

\item{weights}{Optional vector of weights for each observation, or the name of a variable in \code{data} containing the weights.}

\item{B}{A vector of length 2 giving the lower and upper bounds for the observation times. Default is c(0, 1).}

\item{settings}{A list of control settings created by \code{ic_sp_settings()}.}
}
\value{
A list containing the fitted model information, including coefficients, variance-covariance matrix, and other details.
}
\description{
Fit a semi-parametric model for interval-censored data
}
