% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icsp2-package.R
\docType{package}
\name{icsp2-package}
\alias{icsp2}
\alias{icsp2-package}
\title{icsp2: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
