% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icsp2-package.R
\docType{package}
\name{icsp2-package}
\alias{icsp2}
\alias{icsp2-package}
\title{icsp2: Stratified interval-censored semi-parametric survival models}
\description{
A modified implementation of the interval-censored proportional hazards model allowing for stratification, based on icenReg.
}
\seealso{
Useful links:
\itemize{
  \item \url{github.com/igrave/icsp2}
}

}
\author{
\strong{Maintainer}: Isaac Gravestock \email{isaac.gravestock@gmail.com}

Authors:
\itemize{
  \item Clifford Anderson-Bergman
}

}
\keyword{internal}
