% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icsp2-package.R
\docType{package}
\name{icsp2-package}
\alias{icsp2}
\alias{icsp2-package}
\title{icsp2: Stratified interval-censored semi-parametric survival models}
\description{
Semi-parametric survival models for interval censored data with stratification, based on the `icenReg` package, including model fitting and testing.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://igrave.github.io/icsp2/}
  \item Report bugs at \url{github.com/igrave/icsp2}
}

}
\author{
\strong{Maintainer}: Isaac Gravestock \email{isaac.gravestock@gmail.com}

Authors:
\itemize{
  \item Clifford Anderson-Bergman
}

}
\keyword{internal}
