% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icsp2-package.R
\docType{package}
\name{icsp2-package}
\alias{icsp2}
\alias{icsp2-package}
\title{icsp2: Stratified Interval-Censored Semi-Parametric Survival Models}
\description{
Semi-parametric survival models for interval censored data with stratification, based on the 'icenReg' package, including model fitting and testing.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://igrave.github.io/icsp2/}
  \item Report bugs at \url{https://github.com/igrave/icsp2/issues}
}

}
\author{
\strong{Maintainer}: Isaac Gravestock \email{isaac.gravestock@gmail.com} (\href{https://orcid.org/0000-0003-0283-2065}{ORCID})

Authors:
\itemize{
  \item Isaac Gravestock \email{isaac.gravestock@gmail.com} (\href{https://orcid.org/0000-0003-0283-2065}{ORCID})
}

Other contributors:
\itemize{
  \item Clifford Anderson-Bergman (creator of icenReg) [contributor]
}

}
\keyword{internal}
