% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{strata}
\alias{Surv}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{survival}{\code{\link[survival:strata]{strata()}}, \code{\link[survival:Surv]{Surv()}}}
}}

