% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic_sp2.R
\name{vcov.ic_sp2}
\alias{vcov.ic_sp2}
\alias{confint.ic_sp2}
\title{Profile Likelihood Variance and Confidence Intervals for Semi-Parametric Models}
\usage{
\method{vcov}{ic_sp2}(object, type = "oim_curvature", fixed = 5, typical = 1, large = 2, ...)

\method{confint}{ic_sp2}(object, parm, level = 0.95, ...)
}
\arguments{
\item{object}{Fitted model object from \code{ic_sp}}

\item{type}{One of \code{"oim_curvature"} (default), \code{"oim_fixed"}, or \code{"opg_fixed"}. See details for explanation.}

\item{fixed}{A fixed factor to multiply by \code{n^(-1/2)} to determin the perturbation size for fixed types.}

\item{typical}{A typical value for the regression parameters, used to determine the scale of \code{h_n}. Default is 1.
This is required for the \code{"oim_curvature"} type.}

\item{large}{A large value for the regression parameters, used to determine the scale of \code{h_n}. Default is 2.
This is required for the \code{"oim_curvature"} type.}

\item{...}{Unused in \code{vcov()}. In \code{confint()} it is used to specify the vcov parameters such as \code{type}, \code{fixed}, \code{typical}, and \code{large}.}

\item{parm}{a specification of which parameters are to be given confidence intervals, either a vector of numbers or a vector of names. If missing, all parameters are considered.}

\item{level}{the confidence level required.}
}
\value{
Variance-covariance matrix of the regression parameters or the confidence interval matrix.
}
\description{
Profile Likelihood Variance and Confidence Intervals for Semi-Parametric Models
}
\details{
The covariance matrix is calculated using the profile likelihood approach.
(Murphey and Vand Der Vaart 2000).
This method involves perturbing the regression parameters, updating the
baseline hazard estimates using the \code{profile_fit} function with the perturbed
parameters, and calculating the change in log-likelihood for each perturbation,
which is then used to compute the covariance matrix.
We borrowing the naming convention from the Stata \code{stintcox} manual
\url{https://www.stata.com/manuals/ststintcox.pdf}.

Type \code{"oim_curvature"} (Boruvka and Cook 2015) uses the curvature of the log-likelihood function to
determine the perturbation size for the profile likelihood calculations, as
described in Boruvka and Cook (2014).
The \code{typical} and \code{large} parameters are used to determine the scale of the
perturbations.

Type \code{"oim_fixed"} (Zeng et al 2016) uses a fixed perturbation size based on the \code{fixed}
parameter and the sample size \code{n}.

Type \code{"opg_fixed"} (Zeng et al 2017) uses a fixed perturbation size based on the \code{fixed}
parameter and the sample size \code{n}, but uses the outer product of gradients
instead of the observed information matrix for the covariance calculation.

For larged values of \code{fixed} the model fitting may fail to converge.

The \code{confint()} method gives a Wald type CI and is based on \link[stats:confint]{stats::confint} but allows the pass through of the \code{vcov.ic_sp2} parameters.
}
\references{
Murphy, S. A., & Van Der Vaart, A. W. (2000). On Profile Likelihood. Journal of the American Statistical Association, 95(450), 449–465. https://doi.org/10.1080/01621459.2000.10474219

Boruvka, A., and Cook, R. J. (2015), A Cox-Aalen Model for Interval-censored Data. Scand J Statist, 42, 414–426. doi: 10.1111/sjos.12113.

Donglin Zeng, Lu Mao, D. Y. Lin, Maximum likelihood estimation for semiparametric transformation models with interval-censored data, Biometrika, Volume 103, Issue 2, June 2016, Pages 253–271, https://doi.org/10.1093/biomet/asw013

Donglin Zeng, Fei Gao, D. Y. Lin, Maximum likelihood estimation for semiparametric regression models with multivariate interval-censored data, Biometrika, Volume 104, Issue 3, September 2017, Pages 505–525, https://doi.org/10.1093/biomet/asx029
}
